+++ testproject/tests/utils.py
@@ -1,3 +1,5 @@
+from typing import Optional, Tuple
+
 import pytest
 
 from django.conf import settings
@@ -7,6 +9,8 @@ import jwt
 from rest_framework.response import Response
 from rest_framework.test import APIClient
 
+from trench.backends.base import AbstractMessageDispatcher
+from trench.backends.provider import get_mfa_handler
 
 User = get_user_model()
 
@@ -41,3 +45,24 @@ def get_username_from_jwt(response, token_field=default_token_field):
         verify=False,
         algorithms=["HS256"],
     ).get(User.USERNAME_FIELD)
+
+
+def get_authenticated_api_client_and_mfa_handler(user, primary_method: Optional[bool] = None) -> Tuple[APIClient, AbstractMessageDispatcher]:
+    client = APIClient()
+    first_step = login(user)
+    mfa_methods_qs = user.mfa_methods
+    if primary_method is not None:
+        mfa_methods_qs = mfa_methods_qs.filter(is_primary=primary_method)
+    mfa_method = mfa_methods_qs.first()
+    handler = get_mfa_handler(mfa_method=mfa_method)
+    response = client.post(
+        path=PATH_AUTH_JWT_LOGIN_CODE,
+        data={
+            "ephemeral_token": first_step.data.get("ephemeral_token"),
+            "code": handler.create_code(),
+        },
+        format="json",
+    )
+    jwt = get_token_from_response(response)
+    client.credentials(HTTP_AUTHORIZATION=header_template.format(jwt))
+    return client, handler
